+++ src/main/java/com/example/backend/cart/dto/CartCouponResponseDTO.java
@@ -1,5 +1,6 @@
 package com.example.backend.cart.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import lombok.AllArgsConstructor;
@@ -19,8 +20,8 @@ public class CartCouponResponseDTO {
     private Long couponUseId;
     private Long cartId;
     private String couponCode;
-    private Integer discountAmount;
-    private Integer minSpend;
+    private BigDecimal discountAmount;
+    private BigDecimal minSpend;
     private Boolean isActive;
     private LocalDateTime appliedAt;
 }
